refactor: break browse-catalog widget into smaller components
- i18n.ts: translations + useTranslate() hook
- types.ts: shared Product type derived from tool output
- hooks/use-checkout-polling.ts: polling state machine
- components/payment-success.tsx: success card with SVG + order summary
- components/checkout-summary.tsx: order summary + pay/polling button
- components/product-carousel.tsx: product grid, cart toggle, detail panel
- browse-catalog.tsx: thin orchestrator (395 → 134 lines)
